AgentDevelopmentKit_exploration/videomemory/requirements.txt
google-adk
python-dotenv

